Fix edge case where q/qq/Q could start highlight erronerously
Calling a function called `use-for-q` by doing:
`use-for-q @delimiters` would start highlighting because
the @ will match the q_any rules. This is actually totally valid
and we still need to highlight `@` potentially. It's possible
this could also trigger for the other delimiters too, so apply
this to all of them, and only highlight if the previous character
is whitespace or we're at the start of the line.